[testing][rust] Add integration tests, clippy fixes, and dev Makefile (#317)
* Review tech debt: mark completed items, update status tracking

- Silent .catch(() => {}) in column extraction: FIXED (now logs errors)
- Histogram god component: REFACTORED (split into 5 files)
- Docker health checks: partially fixed (clickhouse, db, minio, redis done)
- Docker restart policies: partially fixed (traefik, ingest, backend done)
- Org ClickHouse deletion: marked as skipped (accepted risk)
- Updated line counts to current values


* Mark Rust ingest tests as mostly done in tech debt tracking

Core ingestion routes now have comprehensive integration tests,
auth and stream processor have unit tests. Only files endpoint
integration tests and background processor coverage remain gaps.


* Add integration tests for files endpoint and background processor

Files endpoint tests (5 new):
- Happy path: presigned URL generation with correct grouping
- Multiple files grouped by log_name
- Auth failures: missing auth, wrong key, invalid bearer format
- Missing enrichment headers (X-Project-Name, X-Run-Id)
- Custom file type support

Background processor tests (3 new):
- Inactivity flush: records flushed after interval even below batch_size
- Channel close final flush: buffered records flushed when sender drops
- Batch size flush: records flushed when buffer reaches batch_size

Closes the Rust testing gaps identified in TECH_DEBT.md.


* Add Rust CI job with cargo-llvm-cov coverage and dev Makefile

- Add rust-tests job to CI: fmt check, clippy, tests with LCOV coverage
- Upload coverage artifact for downstream tooling
- Add ingest/Makefile with test, test-cov, test-cov-html, lint, check targets
- Document coverage commands in CLAUDE.md for agent and dev use


* Fix cargo fmt, extract shared test setup, replace sleep with polling

- Run cargo fmt across all Rust source files to fix CI formatting check
- Fix trailing whitespace in stream.rs that blocked rustfmt
- Extract BackgroundProcessorFixture helper struct to reduce code
  duplication across 3 background processor integration tests
- Replace fixed tokio::time::sleep with poll_clickhouse_count that
  retries up to 10s, making tests more robust on slow CI runners


* Fix all clippy warnings to pass CI with -D warnings

- Remove needless borrows for generic args (auth.rs, background.rs)
- Add Default impl for Config (config.rs)
- Use &Path instead of &PathBuf in function signature (cleanup.rs)
- Remove unnecessary u64->u64 casts (replay.rs, main.rs)
- Use .keys() instead of iterating with unused value (metrics.rs)
- Remove .clone() on Copy types (metrics.rs, files.rs)
- Use std::mem::take instead of drain().collect() (stream.rs)
- Suppress dead_code warnings on unused but pub types (status.rs, types.rs, files.rs)
- Allow async_fn_in_trait on internal trait (traits.rs)


* revert: remove ci.yml changes to avoid workflow scope issue

GitHub OAuth App token lacks `workflow` scope needed to push
.github/workflows changes. The Rust CI job will be added separately
via a GitHub-authenticated push.


* Replace sleep-based test waits with polling, fix all clippy uninlined_format_args

Address Gemini review: replace tokio::time::sleep in background processor
tests with poll_clickhouse_count helper for robustness on slow CI runners.
Fix all clippy::uninlined_format_args warnings across ingest crate.


---------

+++ ingest/src/routes/health.rs
@@ -136,7 +136,7 @@ async fn dlq_health(State(state): State<Arc<AppState>>) -> Json<dlq::types::DlqH
             // Actual record counts are available when replaying batches.
             for batch_path in batches {
                 if let Ok(metadata) = tokio::fs::metadata(&batch_path).await {
-                    stats.records_pending += (metadata.len() / 1024) as u64;
+                    stats.records_pending += metadata.len() / 1024;
                 }
             }
         }
